Remove MinimaxNode and remaining gird copy calls.
Only return the score of the state from minimax, this means that the
scores of the children and grandchildren are no longer displayed. Rely
on Zobrist hash to test if two states are equal, this removes the need
to store the grid in the t-table and the remaining calls to the copy
constructor.